clean du code

+++ appli/tft.h
@@ -15,7 +15,7 @@ void TFT_update_info(void);
 void TFT_home_screen(void);
 void TFT_set_vanne(bool_e state);
 void TFT_set_connexion(bool_e state);
-void TFT_set_shower(uint8_t state);
+void TFT_set_shower(uint32_t state);
 void TFT_add_console(char * string);
 
 int16_t * TFT_get_image();
